+++ examples/demo-rollup/stf/src/hooks_impl.rs
@@ -12,9 +12,9 @@ use sov_modules_api::transaction::AuthenticatedTransactionData;
 use sov_modules_api::{
     Context, Gas, ModuleInfo, Spec, StateCheckpoint, StateReaderAndWriter, WorkingSet,
 };
-use sov_modules_stf_blueprint::SequencerOutcome;
+use sov_modules_stf_blueprint::BatchSequencerOutcome;
 use sov_rollup_interface::da::DaSpec;
-use sov_sequencer_registry::SequencerRegistry;
+use sov_sequencer_registry::{SequencerRegistry, SequencerStakeMeter};
 use tracing::info;
 
 use crate::runtime::Runtime;
@@ -25,7 +25,7 @@ impl<S: Spec, Da: DaSpec> TxHooks for Runtime<S, Da> {
 
 impl<S: Spec, Da: DaSpec> ApplyBatchHooks<Da> for Runtime<S, Da> {
     type Spec = S;
-    type BatchResult = SequencerOutcome;
+    type BatchResult = BatchSequencerOutcome;
 
     fn begin_batch_hook(
         &self,
@@ -47,7 +47,7 @@ impl<S: Spec, Da: DaSpec> ApplyBatchHooks<Da> for Runtime<S, Da> {
         // Since we need to make sure the `StfBlueprint` doesn't depend on the module system, we need to
         // convert the `SequencerOutcome` structures manually.
         match result {
-            SequencerOutcome::Rewarded(amount) => {
+            BatchSequencerOutcome::Rewarded(amount) => {
                 info!(%sender, ?amount, "Rewarding sequencer");
                 <SequencerRegistry<S, Da> as ApplyBatchHooks<Da>>::end_batch_hook(
                     &self.sequencer_registry,
@@ -56,8 +56,8 @@ impl<S: Spec, Da: DaSpec> ApplyBatchHooks<Da> for Runtime<S, Da> {
                     state_checkpoint,
                 );
             }
-            SequencerOutcome::Ignored => {}
-            SequencerOutcome::Slashed(reason) => {
+            BatchSequencerOutcome::Ignored => {}
+            BatchSequencerOutcome::Slashed(reason) => {
                 info!(%sender, ?reason, "Slashing sequencer");
                 <SequencerRegistry<S, Da> as ApplyBatchHooks<Da>>::end_batch_hook(
                     &self.sequencer_registry,
@@ -132,7 +132,7 @@ impl<S: Spec, Da: DaSpec> GasEnforcer<S, Da> for Runtime<S, Da> {
         &self,
         tx: &Self::Tx,
         context: &Context<S>,
-        gas_meter: &sov_modules_api::GasMeter<S::Gas>,
+        gas_meter: &sov_modules_api::TxGasMeter<S::Gas>,
         state_checkpoint: &mut StateCheckpoint<S>,
     ) {
         self.bank.refund_remaining_gas(
@@ -147,24 +147,39 @@ impl<S: Spec, Da: DaSpec> GasEnforcer<S, Da> for Runtime<S, Da> {
 }
 
 impl<S: Spec, Da: DaSpec> SequencerAuthorization<S, Da> for Runtime<S, Da> {
+    type SequencerStakeMeter = SequencerStakeMeter<S::Gas>;
+
     fn authorize_sequencer(
         &self,
         sequencer: &<Da as DaSpec>::Address,
+        base_fee_per_gas: &<S::Gas as Gas>::Price,
         state_checkpoint: &mut StateCheckpoint<S>,
-    ) -> Result<(), anyhow::Error> {
+    ) -> Result<SequencerStakeMeter<S::Gas>, anyhow::Error> {
         self.sequencer_registry
-            .authorize_sequencer(sequencer, state_checkpoint)
+            .authorize_sequencer(sequencer, base_fee_per_gas, state_checkpoint)
             .context("An error occurred while checking the sequencer bond")
+    }
+
+    fn refund_sequencer(
+        &self,
+        sequencer_stake_meter: &mut Self::SequencerStakeMeter,
+        refund_amount: u64,
+    ) {
+        self.sequencer_registry
+            .refund_sequencer(sequencer_stake_meter, refund_amount);
     }
 
     fn penalize_sequencer(
         &self,
         sequencer: &Da::Address,
-        amount: u64,
+        sequencer_stake_meter: SequencerStakeMeter<S::Gas>,
         state_checkpoint: &mut StateCheckpoint<S>,
     ) {
-        self.sequencer_registry
-            .penalize_sequencer(sequencer, amount, state_checkpoint);
+        self.sequencer_registry.penalize_sequencer(
+            sequencer,
+            sequencer_stake_meter,
+            state_checkpoint,
+        );
     }
 }