+++ examples/demo-rollup/stf/src/tests/stf_tests.rs
@@ -3,7 +3,7 @@ use std::vec;
 use sov_mock_da::{MockBlock, MockDaSpec, MOCK_SEQUENCER_DA_ADDRESS};
 use sov_modules_api::batch::BatchWithId;
 use sov_modules_api::{Spec, WorkingSet};
-use sov_modules_stf_blueprint::{SequencerOutcome, StfBlueprint};
+use sov_modules_stf_blueprint::{BatchSequencerOutcome, StfBlueprint};
 use sov_rollup_interface::da::RelevantBlobs;
 use sov_rollup_interface::services::da::SlotData;
 use sov_rollup_interface::stf::StateTransitionFunction;
@@ -71,7 +71,7 @@ fn test_demo_values_in_db() {
 
         let apply_blob_outcome = result.batch_receipts[0].clone();
         assert_eq!(
-            SequencerOutcome::Rewarded(0),
+            BatchSequencerOutcome::Rewarded(0),
             apply_blob_outcome.inner,
             "Sequencer execution should have succeeded but failed "
         );
@@ -159,7 +159,7 @@ fn test_demo_values_in_cache() {
     let apply_blob_outcome = apply_block_result.batch_receipts[0].clone();
 
     assert_eq!(
-        SequencerOutcome::Rewarded(0),
+        BatchSequencerOutcome::Rewarded(0),
         apply_blob_outcome.inner,
         "Sequencer execution should have succeeded but failed"
     );