+++ module-system/module-implementations/sov-attester-incentives/src/tests/attestation_processing.rs
@@ -1,7 +1,7 @@
 use sov_bank::GAS_TOKEN_ID;
 use sov_modules_api::optimistic::Attestation;
 use sov_modules_api::{Context, WorkingSet};
-use sov_modules_core::GasMeter;
+use sov_modules_core::TxGasMeter;
 use sov_prover_storage_manager::new_orphan_storage;
 
 use crate::call::AttesterIncentiveErrors;
@@ -45,7 +45,7 @@ fn test_process_valid_attestation() {
     let transition_1 = exec_vars.pop().unwrap();
     let initial_transition = exec_vars.pop().unwrap();
 
-    let mut working_set = state_checkpoint.to_revertable(GasMeter::unmetered());
+    let mut working_set = state_checkpoint.to_revertable(TxGasMeter::unmetered());
 
     // Process a valid attestation for the first transition
     {
@@ -150,7 +150,7 @@ fn test_burn_on_invalid_attestation() {
 
     let context = Context::<S>::new(attester_address, sequencer, 1);
 
-    let mut working_set = state_checkpoint.to_revertable(GasMeter::unmetered());
+    let mut working_set = state_checkpoint.to_revertable(TxGasMeter::unmetered());
     // Process an invalid proof for genesis: everything is correct except the storage proof.
     // Must simply return an error. Cannot burn the token at this point because we don't know if the
     // sender is bonded or not.